refactored filtering code to separate module
